requirements.txt
torch
torchvision
efficientnet_pytorch
flask
Pillow
numpy
